separate api server args into another file
